refactor: fetch unico para json mockado
