+++ src/components/Footer.js
@@ -4,7 +4,7 @@ export const Footer = () => {
   return (
     <>
     <footer className="bg-cyan-600 px-10 py-5 font-inter">
-        <p className="text-white text-center">Desenvolvido por Caio P., João K., Erich J., Renato A., Sarah C.</p>
+        <p className="text-white text-center">Desenvolvido por Caio P., João K., Erich J., Renato A. e Sarah C.</p>
     </footer>
     </>
